moved auth struct to /models/auth.go
